Refactored code: removed redudant comments and moved code to nc_util.c
Removed redudant files
